requirements.txt
Django==3.1.12
djangorestframework==3.12.2
django-allauth==0.44.0

psycopg2-binary==2.8.6

uwsgi==2.0.19.1

django-redis==4.12.1
celery==4.4.7

drf-yasg==1.20.0

Pillow==8.2.0
django-imagekit

markdown
django-filter

filetype
django-mptt

django-crispy-forms

requests==2.25.0

django-celery-email
m3u8

django-ckeditor
django-debug-toolbar

django-login-required-middleware==0.6.1
